feat: Record more mining info
