+++ miner/api.go
@@ -27,24 +27,3 @@ type IMinerManager interface {
 	StatesForMining(context.Context, []address.Address) ([]types.MinerState, error)
 	CountWinners(context.Context, []address.Address, abi.ChainEpoch, abi.ChainEpoch) ([]types.CountWinners, error)
 }
-
-type MockMinerMgr struct {
-}
-
-var _ IMinerManager = &MockMinerMgr{}
-
-func (m MockMinerMgr) UpdateAddress(context.Context, int64, int64) ([]types.MinerInfo, error) {
-	return nil, nil
-}
-
-func (m MockMinerMgr) ListAddress(context.Context) ([]types.MinerInfo, error) {
-	return nil, nil
-}
-
-func (m MockMinerMgr) StatesForMining(context.Context, []address.Address) ([]types.MinerState, error) {
-	return nil, nil
-}
-
-func (m MockMinerMgr) CountWinners(context.Context, []address.Address, abi.ChainEpoch, abi.ChainEpoch) ([]types.CountWinners, error) {
-	return nil, nil
-}
