+++ node/modules/slashfilter/api.go
@@ -2,6 +2,7 @@ package slashfilter
 
 import (
 	"context"
+	"time"
 
 	"github.com/filecoin-project/go-state-types/abi"
 
@@ -15,8 +16,18 @@ const (
 	MySQL BlockStoreType = "mysql"
 )
 
+type StateMining int
+
+const (
+	Mining StateMining = iota
+	Success
+	Timeout
+	ChainForked
+	Error
+)
+
 type SlashFilterAPI interface {
 	HasBlock(ctx context.Context, bh *vtypes.BlockHeader) (bool, error)
 	MinedBlock(ctx context.Context, bh *vtypes.BlockHeader, parentEpoch abi.ChainEpoch) error
-	PutBlock(ctx context.Context, bh *vtypes.BlockHeader, parentEpoch abi.ChainEpoch) error
+	PutBlock(ctx context.Context, bh *vtypes.BlockHeader, parentEpoch abi.ChainEpoch, t time.Time, state StateMining) error
 }
